requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-limiter==3.9.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
pandas==2.2.3
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
waitress==3.0.2
Werkzeug==3.1.3